API done

+++ functions/APIs/todos.js
@@ -2,6 +2,7 @@ const { db } = require("../util/admin");
 
 exports.getAllTodos = (request, response) => {
   db.collection("todos")
+    .where("username", "==", request.user.username)
     .orderBy("createdAt", "desc")
     .get()
     .then((data) => {
@@ -22,6 +23,20 @@ exports.getAllTodos = (request, response) => {
     });
 };
 
+exports.getOneTodo = (request, response) => {
+  db.doc(`/todos/${request.params.todoId}`)
+    .get()
+    .then((doc) => {
+      if (!doc.exists) {
+        return response.status(404).json({ error: "Todo not found" });
+      }
+      return response.json(doc.data());
+    })
+    .catch((err) => {
+      console.error(err);
+      return response.status(500).json({ error: err.code });
+    });
+};
 exports.postOneTodo = (request, response) => {
   if (request.body.body.trim() === "") {
     return response.status(400).json({ body: "Can not be empty" });
@@ -34,6 +49,7 @@ exports.postOneTodo = (request, response) => {
     title: request.body.title,
     body: request.body.body,
     createdAt: new Date().toISOString(),
+    username: request.user.username,
   };
 
   db.collection("todos")
@@ -58,6 +74,9 @@ exports.deleteTodo = (request, response) => {
       if (!doc.exists) {
         return response.status(404).json({ error: "Todo not found" });
       }
+      if (doc.data().username !== request.user.username) {
+        return response.status(403).json({ error: "Unauthorized" });
+      }
       return document.delete();
     })
     .then(() => {
@@ -75,6 +94,16 @@ exports.editTodo = (request, response) => {
   }
 
   let document = db.collection("/todos").doc(`${request.params.todoId}`);
+  document
+    .get()
+    .then((doc) => {
+      if (doc.data().username !== request.body.username)
+        return response.status(403).json({ error: "Unauthorized" });
+    })
+    .catch((err) => {
+      console.error(err);
+      return response.status(500).json({ error: err.code });
+    });
   document
     .update(request.body)
     .then(() => response.json({ message: "Update successfully" }))
